requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
requests==2.31.0
